+++ webpack.config.js
@@ -8,11 +8,6 @@ const coreConfig = () => {
 
   const i18nAlias = path.resolve(path.join(rootPath, 'packages', 'terra-i18n'));
   return {
-    resolveLoader: {
-      alias: {
-        'terra-props-table-loader': path.join(__dirname, 'scripts/terra-props-table-loader'),
-      },
-    },
     resolve: {
       alias: {
         'terra-i18n': i18nAlias,